Added Dash cooldown bar and level up visualization

+++ Hells Gate/Assets/Weapon/Skill/skill1.cs
@@ -16,7 +16,7 @@ public class skill1 : SkillParent
         if (collision.gameObject.CompareTag("Enemy"))
         {
             collision.GetComponent<enemy>()
-            .takeDmg(damage);
+            .takeDmg((int)damage);
             Debug.Log("111111111111");
         }
